modificando la relación de las tablas y los controladores

+++ src/main/java/com/upeu/demo/controller/CarreraController.java
@@ -1,16 +1,14 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.upeu.demo.controller;
 
 import com.upeu.demo.entity.Carrera;
 import com.upeu.demo.service.CarreraService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import java.util.List;
+import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@CrossOrigin(origins = {"http://localhost:4200"})/*Hare peticiones a tal ruta externa*/
 @RestController
 @RequestMapping("/api/carrera")
 @Api(value = "Microservicios de gestion de carreras", description ="Microservicio de carreras")
@@ -28,10 +27,14 @@ public class CarreraController {
     @Autowired
     private CarreraService carreraService;
 
-    @ApiOperation(value="Lista de carrera")
-    @GetMapping("/all")
-    public List<Carrera> findAll() {
-        return carreraService.findAll();
+    @ApiOperation(value="Lista de carreras")
+    @GetMapping()
+    public ResponseEntity<?> findAll() {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Lista de carreras");
+        result.put("data", carreraService.findAll());
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Obtiene datos de la carrera")
@@ -43,19 +46,51 @@ public class CarreraController {
 
     @ApiOperation(value="Elimina una carrera")
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        carreraService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Carrera data = carreraService.findById(id);
+        if(data == null){
+            result.put("success", false);
+            result.put("message", "No existe carrera con id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }else{
+            carreraService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Carrera eliminado correctamente");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value="Crea una carrera")
-    @PostMapping("/save")
-    public Carrera save(@RequestBody Carrera carrera) {
-        return carreraService.save(carrera);
+    @PostMapping
+    public ResponseEntity<?> save(@RequestBody Carrera carrera) {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Carrera registrado correctamente");
+        result.put("data", carreraService.save(carrera));
+
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Modifica una carrera")
-    @PutMapping("/update")
-    public Carrera update(@RequestBody Carrera carrera) {
-        return carreraService.save(carrera);
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Carrera carrera) {
+                HashMap<String, Object> result = new HashMap<>();
+        Carrera data = carreraService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe registro con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            carrera.setCarrId(id);
+            carreraService.save(carrera);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", carrera);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
