+++ src/main/java/com/upeu/demo/controller/MaterialController.java
@@ -1,16 +1,14 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.upeu.demo.controller;
 
 import com.upeu.demo.entity.Material;
 import com.upeu.demo.service.MaterialService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import java.util.List;
+import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@CrossOrigin(origins = {"http://localhost:4200"})/*Hare peticiones a tal ruta externa*/
 @RestController
 @RequestMapping("/api/material")
 @Api(value = "Microservicios de gestion de material", description ="Microservicio de material")
@@ -29,9 +28,13 @@ public class MaterialController {
     private MaterialService materialService;
 
     @ApiOperation(value="Lista de material")
-    @GetMapping("/all")
-    public List<Material> findAll() {
-        return materialService.findAll();
+    @GetMapping()
+    public ResponseEntity<?> findAll() {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Lista de programas");
+        result.put("data", materialService.findAll());
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Obtiene datos de material")
@@ -43,19 +46,51 @@ public class MaterialController {
 
     @ApiOperation(value="Elimina un material")
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        materialService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Material data = materialService.findById(id);
+        if(data == null){
+            result.put("success", false);
+            result.put("message", "No existe material con id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }else{
+            materialService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Material eliminado correctamente");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value="Crea un material")
-    @PostMapping("/save")
-    public Material save(@RequestBody Material material) {
-        return materialService.save(material);
+    @PostMapping
+    public ResponseEntity<?> save(@RequestBody Material material) {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Material registrado correctamente");
+        result.put("data", materialService.save(material));
+
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Modifica un material")
-    @PutMapping("/update")
-    public Material update(@RequestBody Material material) {
-        return materialService.save(material);
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Material material) {
+        HashMap<String, Object> result = new HashMap<>();
+        Material data = materialService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe registro con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            material.setMateId(id);
+            materialService.save(material);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", material);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
