+++ src/main/java/com/upeu/demo/controller/ProgramaController.java
@@ -8,9 +8,11 @@ import com.upeu.demo.entity.Programa;
 import com.upeu.demo.service.ProgramaService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import java.util.List;
+import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +22,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@CrossOrigin(origins = {"http://localhost:4200"})/*Hare peticiones a tal ruta externa*/
 @RestController
 @RequestMapping("/api/programa")
 @Api(value = "Microservicios de gestion de programa", description ="Microservicio de programa")
@@ -28,10 +31,14 @@ public class ProgramaController {
     @Autowired
     private ProgramaService programaService;
 
-    @ApiOperation(value="Lista de programa")
-    @GetMapping("/all")
-    public List<Programa> findAll() {
-        return programaService.findAll();
+    @ApiOperation(value="Lista de programas")
+    @GetMapping()
+    public ResponseEntity<?> findAll() {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Lista de programas");
+        result.put("data", programaService.findAll());
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Obtiene datos de programa")
@@ -43,19 +50,51 @@ public class ProgramaController {
 
     @ApiOperation(value="Elimina un programa")
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        programaService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if(data == null){
+            result.put("success", false);
+            result.put("message", "No existe programa con id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }else{
+            programaService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Programa eliminado correctamente");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value="Crea un programa")
-    @PostMapping("/save")
-    public Programa save(@RequestBody Programa programa) {
-        return programaService.save(programa);
+    @PostMapping
+    public ResponseEntity<?> save(@RequestBody Programa programa){
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Programa registrado correctamente");
+        result.put("data", programaService.save(programa));
+
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
-    @ApiOperation(value="Modifica un programa")
-    @PutMapping("/update")
-    public Programa update(@RequestBody Programa programa) {
-        return programaService.save(programa);
+    @ApiOperation(value = "Modifica una programa")
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Programa programa) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe registro con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            programa.setProgId(id);
+            programaService.save(programa);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", programa);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
