+++ src/main/java/com/upeu/demo/controller/TallerController.java
@@ -4,9 +4,11 @@ import com.upeu.demo.entity.Taller;
 import com.upeu.demo.service.TallerService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import java.util.List;
+import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@CrossOrigin(origins = {"http://localhost:4200"})/*Hare peticiones a tal ruta externa*/
 @RestController
 @RequestMapping("/api/taller")
 @Api(value = "Microservicios de gestion de taller", description ="Microservicio de taller")
@@ -25,9 +28,13 @@ public class TallerController {
     private TallerService tallerService;
 
     @ApiOperation(value="Lista de taller")
-    @GetMapping("/all")
-    public List<Taller> findAll() {
-        return tallerService.findAll();
+    @GetMapping()
+    public ResponseEntity<?> findAll() {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Lista de taller");
+        result.put("data", tallerService.findAll());
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Obtiene datos de taller")
@@ -39,19 +46,51 @@ public class TallerController {
 
     @ApiOperation(value="Elimina un taller")
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        tallerService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Taller data = tallerService.findById(id);
+        if(data == null){
+            result.put("success", false);
+            result.put("message", "No existe taller con id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }else{
+            tallerService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Taller eliminado correctamente");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value="Crea un taller")
-    @PostMapping("/save")
-    public Taller save(@RequestBody Taller taller) {
-        return tallerService.save(taller);
+    @PostMapping
+    public ResponseEntity<?> save(@RequestBody Taller taller) {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Taller registrado correctamente");
+        result.put("data", tallerService.save(taller));
+
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @ApiOperation(value="Modifica un taller")
-    @PutMapping("/update")
-    public Taller update(@RequestBody Taller taller) {
-        return tallerService.save(taller);
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Taller taller) {
+        HashMap<String, Object> result = new HashMap<>();
+        Taller data = tallerService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe registro con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            taller.setTallId(id);
+            tallerService.save(taller);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", taller);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
